# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Procurement System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  fullName: string;
  role: string;
  department: string;
  /** @nullable */
  approvalStage: number | null;
  isActive: boolean;
}

export interface User {
  id: number;
  username: string;
  fullName: string;
  role: string;
  department: string;
  /** @nullable */
  approvalStage: number | null;
  isActive: boolean;
  createdAt: string;
}

export interface UserInput {
  username: string;
  password: string;
  fullName: string;
  role: string;
  department: string;
  /** @nullable */
  approvalStage?: number | null;
}

export interface UserUpdate {
  fullName?: string;
  role?: string;
  department?: string;
  /** @nullable */
  approvalStage?: number | null;
  isActive?: boolean;
  password?: string;
}

export interface LineItemInput {
  description: string;
  quantity: number;
  unitPrice: number;
  category: string;
}

export interface LineItem {
  id: number;
  description: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  category: string;
}

export interface ApprovalRecord {
  id: number;
  approverName: string;
  stage: number;
  action: string;
  /** @nullable */
  comments?: string | null;
  createdAt: string;
}

export interface RequisitionInput {
  title: string;
  description: string;
  department: string;
  isUrgent?: boolean;
  justification: string;
  lineItems: LineItemInput[];
}

export interface RequisitionSummary {
  id: number;
  prfNumber: string;
  title: string;
  department: string;
  /** @nullable */
  currentStage: number | null;
  status: string;
  isUrgent: boolean;
  totalAmount: number;
  createdAt: string;
  creatorName: string;
}

export interface RequisitionDetail {
  id: number;
  prfNumber: string;
  title: string;
  description: string;
  department: string;
  /** @nullable */
  currentStage: number | null;
  status: string;
  isUrgent: boolean;
  justification: string;
  totalAmount: number;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  completedAt?: string | null;
  creatorId?: number;
  creatorName: string;
  lineItems: LineItem[];
  approvalHistory: ApprovalRecord[];
}

export interface ApprovalInput {
  comments?: string;
}

export interface Notification {
  id: number;
  title: string;
  message: string;
  notificationType: string;
  isRead: boolean;
  /** @nullable */
  relatedRequisitionId?: number | null;
  createdAt: string;
}

export interface DashboardStats {
  total: number;
  draft: number;
  pending: number;
  approved: number;
  rejected: number;
  totalAmount: number;
  unreadNotifications?: number;
}

export interface StageCount {
  stage: number;
  stageName: string;
  count: number;
}

export type ListRequisitionsParams = {
status?: string;
/**
 * @nullable
 */
stage?: number | null;
};