# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Procurement System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requisitions:
    get:
      operationId: listRequisitions
      tags: [requisitions]
      summary: List requisitions (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of requisitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequisitionSummary"
    post:
      operationId: createRequisition
      tags: [requisitions]
      summary: Create a new requisition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
  /requisitions/{id}:
    get:
      operationId: getRequisition
      tags: [requisitions]
      summary: Get requisition detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Requisition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requisitions/{id}/submit:
    post:
      operationId: submitRequisition
      tags: [requisitions]
      summary: Submit requisition for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "400":
          description: Cannot submit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requisitions/{id}/approve:
    post:
      operationId: approveRequisition
      tags: [requisitions]
      summary: Approve a requisition at current stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requisitions/{id}/reject:
    post:
      operationId: rejectRequisition
      tags: [requisitions]
      summary: Reject a requisition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequisitionDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, fullName, role, department, approvalStage, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        department:
          type: string
        approvalStage:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    LineItemInput:
      type: object
      required: [description, quantity, unitPrice, category]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        category:
          type: string
    LineItem:
      type: object
      required: [id, description, quantity, unitPrice, totalPrice, category]
      properties:
        id:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
        category:
          type: string
    ApprovalRecord:
      type: object
      required: [id, approverName, stage, action, createdAt]
      properties:
        id:
          type: integer
        approverName:
          type: string
        stage:
          type: integer
        action:
          type: string
        comments:
          type: ["string", "null"]
        createdAt:
          type: string
    RequisitionInput:
      type: object
      required: [title, description, department, justification, lineItems]
      properties:
        title:
          type: string
        description:
          type: string
        department:
          type: string
        isUrgent:
          type: boolean
        justification:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
    RequisitionSummary:
      type: object
      required: [id, prfNumber, title, department, currentStage, status, isUrgent, totalAmount, createdAt, creatorName]
      properties:
        id:
          type: integer
        prfNumber:
          type: string
        title:
          type: string
        department:
          type: string
        currentStage:
          type: ["integer", "null"]
        status:
          type: string
        isUrgent:
          type: boolean
        totalAmount:
          type: number
        createdAt:
          type: string
        creatorName:
          type: string
    RequisitionDetail:
      type: object
      required: [id, prfNumber, title, description, department, currentStage, status, isUrgent, justification, totalAmount, createdAt, creatorName, lineItems, approvalHistory]
      properties:
        id:
          type: integer
        prfNumber:
          type: string
        title:
          type: string
        description:
          type: string
        department:
          type: string
        currentStage:
          type: ["integer", "null"]
        status:
          type: string
        isUrgent:
          type: boolean
        justification:
          type: string
        totalAmount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        creatorId:
          type: integer
        creatorName:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        approvalHistory:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRecord"
    ApprovalInput:
      type: object
      properties:
        comments:
          type: string
...