# lib/api-zod/src/generated/types/approvalRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Procurement System API
 * OpenAPI spec version: 0.1.0
 */

export interface ApprovalRecord {
  id: number;
  approverName: string;
  stage: number;
  action: string;
  /** @nullable */
  comments?: string | null;
  createdAt: string;
}