# lib/api-zod/src/generated/types/requisitionDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Procurement System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalRecord } from './approvalRecord';
import type { LineItem } from './lineItem';

export interface RequisitionDetail {
  id: number;
  prfNumber: string;
  title: string;
  description: string;
  department: string;
  /** @nullable */
  currentStage: number | null;
  status: string;
  isUrgent: boolean;
  justification: string;
  totalAmount: number;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  completedAt?: string | null;
  creatorId?: number;
  creatorName: string;
  lineItems: LineItem[];
  approvalHistory: ApprovalRecord[];
}