# lib/api-zod/src/generated/types/requisitionSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * E-Procurement System API
 * OpenAPI spec version: 0.1.0
 */

export interface RequisitionSummary {
  id: number;
  prfNumber: string;
  title: string;
  department: string;
  /** @nullable */
  currentStage: number | null;
  status: string;
  isUrgent: boolean;
  totalAmount: number;
  createdAt: string;
  creatorName: string;
}